+++ waldo/data/global.go
@@ -3,5 +3,7 @@ package data
 const (
 	CLIName    = "Waldo CLI"
 	CLIPrefix  = "waldo"
-	CLIVersion = "3.2.1"
+	CLIVersion = "3.3.0"
+
+	CoreAPIUserEndpoint = "https://api.waldo.com/1.0/users/me"
 )