+++ CHANGELOG.md
@@ -7,6 +7,16 @@ This project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [3.3.0] - 2024-03-08
+
+### Added
+
+- Added new `auth` verb to support authorizing with a user token.
+
+### Changed
+
+- Upgraded all dependencies to latest and greatest.
+
 ## [3.2.1] - 2023-12-14
 
 ### Changed
@@ -129,7 +139,8 @@ Initial public release of enhanced functionality.
 
 Initial public release.
 
-[Unreleased]:   https://github.com/waldoapp/waldo-go-cli/compare/3.2.1...HEAD
+[Unreleased]:   https://github.com/waldoapp/waldo-go-cli/compare/3.3.0...HEAD
+[3.3.0]:		https://github.com/waldoapp/waldo-go-cli/compare/3.2.1...3.3.0
 [3.2.1]:		https://github.com/waldoapp/waldo-go-cli/compare/3.2.0...3.2.1
 [3.2.0]:		https://github.com/waldoapp/waldo-go-cli/compare/3.1.0...3.2.0
 [3.1.0]:		https://github.com/waldoapp/waldo-go-cli/compare/3.0.0...3.1.0